+++ x/posts/internal/keeper/handler_test.go
@@ -2,7 +2,6 @@ package keeper_test
 
 import (
 	"fmt"
-	"strconv"
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -27,28 +26,28 @@ func Test_handleMsgCreatePost(t *testing.T) {
 		{
 			name: "Trying to store post with same id returns expError",
 			storedPosts: types.Posts{
-				types.NewPost(types.PostID(1), testPost.ParentID, testPost.Message, testPost.AllowsComments, "desmos", map[string]string{}, testPost.Created.Int64(), testPost.Creator),
+				types.NewPost(types.PostID(1), testPost.ParentID, testPost.Message, testPost.AllowsComments, "desmos", map[string]string{}, testPost.Created, testPost.Creator),
 			},
 			lastPostID: types.PostID(0),
-			msg:        types.NewMsgCreatePost(testPost.Message, testPost.ParentID, testPost.AllowsComments, "desmos", map[string]string{}, testPost.Creator),
+			msg:        types.NewMsgCreatePost(testPost.Message, testPost.ParentID, testPost.AllowsComments, "desmos", map[string]string{}, testPost.Creator, testPost.Created),
 			expError:   "Post with id 1 already exists",
 		},
 		{
 			name:    "Post with new id is stored properly",
-			msg:     types.NewMsgCreatePost(testPost.Message, testPost.ParentID, false, "desmos", map[string]string{}, testPost.Creator),
-			expPost: types.NewPost(types.PostID(1), testPost.ParentID, testPost.Message, testPost.AllowsComments, "desmos", map[string]string{}, 0, testPost.Creator),
+			msg:     types.NewMsgCreatePost(testPost.Message, testPost.ParentID, testPost.AllowsComments, testPost.Subspace, testPost.OptionalData, testPost.Creator, testPost.Created),
+			expPost: types.NewPost(types.PostID(1), testPost.ParentID, testPost.Message, testPost.AllowsComments, testPost.Subspace, testPost.OptionalData, testPost.Created, testPost.Creator),
 		},
 		{
 			name:     "Storing a valid post with missing parent id returns expError",
-			msg:      types.NewMsgCreatePost(testPost.Message, types.PostID(50), false, "desmos", map[string]string{}, testPost.Creator),
+			msg:      types.NewMsgCreatePost(testPost.Message, types.PostID(50), false, "desmos", map[string]string{}, testPost.Creator, testPost.Created),
 			expError: "Parent post with id 50 not found",
 		},
 		{
 			name: "Storing a valid post with parent stored but not accepting comments returns expError",
 			storedPosts: types.Posts{
-				types.NewPost(types.PostID(50), types.PostID(50), "Parent post", false, "desmos", map[string]string{}, 0, testPost.Creator),
+				types.NewPost(types.PostID(50), types.PostID(50), "Parent post", false, "desmos", map[string]string{}, testPost.Created, testPost.Creator),
 			},
-			msg:      types.NewMsgCreatePost(testPost.Message, types.PostID(50), false, "desmos", map[string]string{}, testPost.Creator),
+			msg:      types.NewMsgCreatePost(testPost.Message, types.PostID(50), false, "desmos", map[string]string{}, testPost.Creator, testPost.Created),
 			expError: "Post with id 50 does not allow comments",
 		},
 	}
@@ -77,7 +76,7 @@ func Test_handleMsgCreatePost(t *testing.T) {
 				// Check the post
 				var stored types.Post
 				k.Cdc.MustUnmarshalBinaryBare(store.Get([]byte(types.PostStorePrefix+test.expPost.PostID.String())), &stored)
-				assert.True(t, stored.Equals(test.expPost))
+				assert.True(t, stored.Equals(test.expPost), "Expected: %s, actual: %s", test.expPost, stored)
 
 				// Check the data
 				assert.Equal(t, k.Cdc.MustMarshalBinaryLengthPrefixed(test.expPost.PostID), res.Data)
@@ -106,44 +105,42 @@ func Test_handleMsgCreatePost(t *testing.T) {
 
 func Test_handleMsgEditPost(t *testing.T) {
 	editor, _ := sdk.AccAddressFromBech32("cosmos1z427v6xdc8jgn5yznfzhwuvetpzzcnusut3z63")
+
 	testData := []struct {
-		name        string
-		storedPost  *types.Post
-		msg         types.MsgEditPost
-		blockHeight int64
-		expError    string
-		expPost     types.Post
+		name       string
+		storedPost *types.Post
+		msg        types.MsgEditPost
+		expError   string
+		expPost    types.Post
 	}{
 		{
 			name:       "Post not found",
 			storedPost: nil,
-			msg:        types.NewMsgEditPost(types.PostID(0), "Edited message", testPostOwner),
+			msg:        types.NewMsgEditPost(types.PostID(0), "Edited message", testPostOwner, testPost.Created),
 			expError:   "Post with id 0 not found",
 		},
 		{
 			name:       "Invalid editor",
 			storedPost: &testPost,
-			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", editor),
+			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", editor, testPost.Created),
 			expError:   "Incorrect owner",
 		},
 		{
-			name:        "Edit date before creation date",
-			storedPost:  &testPost,
-			blockHeight: testPost.Created.Int64() - 1,
-			msg:         types.NewMsgEditPost(testPost.PostID, "Edited message", testPost.Creator),
-			expError:    "Edit date cannot be before creation date",
+			name:       "Edit date before creation date",
+			storedPost: &testPost,
+			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", testPost.Creator, testPost.Created.AddDate(0, 0, -1)),
+			expError:   "Edit date cannot be before creation date",
 		},
 		{
-			name:        "Valid request is handled properly",
-			storedPost:  &testPost,
-			blockHeight: testPost.Created.Int64() + 1,
-			msg:         types.NewMsgEditPost(testPost.PostID, "Edited message", testPost.Creator),
+			name:       "Valid request is handled properly",
+			storedPost: &testPost,
+			msg:        types.NewMsgEditPost(testPost.PostID, "Edited message", testPost.Creator, testPost.Created.AddDate(0, 0, 1)),
 			expPost: types.Post{
 				PostID:         testPost.PostID,
 				ParentID:       testPost.ParentID,
 				Message:        "Edited message",
 				Created:        testPost.Created,
-				LastEdited:     testPost.Created.Add(sdk.NewInt(1)),
+				LastEdited:     testPost.Created.AddDate(0, 0, 1),
 				AllowsComments: testPost.AllowsComments,
 				Subspace:       testPost.Subspace,
 				OptionalData:   testPost.OptionalData,
@@ -156,9 +153,6 @@ func Test_handleMsgEditPost(t *testing.T) {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
 			ctx, k := SetupTestInput()
-			if test.blockHeight != 0 {
-				ctx = ctx.WithBlockHeight(test.blockHeight)
-			}
 
 			store := ctx.KVStore(k.StoreKey)
 			if test.storedPost != nil {
@@ -176,8 +170,8 @@ func Test_handleMsgEditPost(t *testing.T) {
 				assert.True(t, res.IsOK())
 				assert.Contains(t, res.Events, sdk.NewEvent(
 					types.EventTypePostEdited,
-					sdk.NewAttribute(types.AttributeKeyPostID, testPost.PostID.String()),
-					sdk.NewAttribute(types.AttributeKeyPostEditTime, strconv.FormatInt(ctx.BlockHeight(), 10)),
+					sdk.NewAttribute(types.AttributeKeyPostID, test.msg.PostID.String()),
+					sdk.NewAttribute(types.AttributeKeyPostEditTime, test.msg.EditDate.String()),
 				))
 
 				var stored types.Post
@@ -221,9 +215,6 @@ func Test_handleMsgAddPostReaction(t *testing.T) {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
 			ctx, k := SetupTestInput()
-			if test.existingPost != nil {
-				ctx = ctx.WithBlockHeight(test.existingPost.Created.Int64() + 1)
-			}
 
 			store := ctx.KVStore(k.StoreKey)
 			if test.existingPost != nil {
@@ -252,7 +243,7 @@ func Test_handleMsgAddPostReaction(t *testing.T) {
 
 				var storedReactions types.Reactions
 				k.Cdc.MustUnmarshalBinaryBare(store.Get([]byte(types.PostReactionsStorePrefix+storedPost.PostID.String())), &storedReactions)
-				assert.Contains(t, storedReactions, types.NewReaction(test.msg.Value, ctx.BlockHeight(), test.msg.User))
+				assert.Contains(t, storedReactions, types.NewReaction(test.msg.Value, test.msg.User))
 			}
 
 			// Invalid response
@@ -266,7 +257,7 @@ func Test_handleMsgAddPostReaction(t *testing.T) {
 
 func Test_handleMsgRemovePostReaction(t *testing.T) {
 	user, _ := sdk.AccAddressFromBech32("cosmos1q4hx350dh0843wr3csctxr87at3zcvd9qehqvg")
-	reaction := types.NewReaction("like", testPost.Created.Int64()+1, user)
+	reaction := types.NewReaction("like", user)
 	tests := []struct {
 		name             string
 		existingPost     *types.Post
@@ -298,9 +289,6 @@ func Test_handleMsgRemovePostReaction(t *testing.T) {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
 			ctx, k := SetupTestInput()
-			if test.existingPost != nil {
-				ctx = ctx.WithBlockHeight(test.existingPost.Created.Int64() + 1)
-			}
 
 			store := ctx.KVStore(k.StoreKey)
 			if test.existingPost != nil {