+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implemented the support for posts reactions (#47)
 - Implemented the support for posts subspaces (#46)
 - Automated the default bond denom change to `desmos` (#25)
+- Replaced the block height with timestamps inside posts' creation dates and edit dates (#62)
 - Capped the post message length to 500 characters (#67)
 
 ## Migration